+++ runtime/rococo/src/migration/mod.rs
@@ -12,11 +12,11 @@ pub mod parachain_staking;
 pub use parachain_staking::ReplaceParachainStakingStorage;
 pub mod balances_transaction_payment;
 pub use balances_transaction_payment::ReplaceBalancesRelatedStorage;
-pub mod pallet_bounty;
-pub use pallet_bounty::ReplacePalletBountyStorage;
-pub mod pallet_democracy;
-pub use pallet_democracy::ReplaceDemocracyStorage;
-pub mod pallet_preimage;
-pub use pallet_preimage::ReplacePreImageStorage;
-pub mod pallet_treasury;
-pub use pallet_treasury::ReplaceTreasuryStorage;
+pub mod migrate_bounty;
+pub use migrate_bounty::ReplacePalletBountyStorage;
+pub mod migrate_democracy;
+pub use migrate_democracy::ReplaceDemocracyStorage;
+pub mod migrate_preimage;
+pub use migrate_preimage::ReplacePreImageStorage;
+pub mod migrate_treasury;
+pub use migrate_treasury::ReplaceTreasuryStorage;
